Move qr code to component; include game name
