+++ src/games/farsketched/__tests__/reducer.test.ts
@@ -87,7 +87,7 @@ describe('farsketchedReducer', () => {
 
   it('should handle connection messages without state changes', () => {
     const connectionMessages = [
-      createMessage<ConnectionRequestMessage>(MessageType.CONNECTION_REQUEST, { playerId: 'test', roomCode: 'test' }),
+      createMessage<ConnectionRequestMessage>(MessageType.CONNECTION_REQUEST, { playerId: 'test', room: 'test' }),
       createMessage<ConnectionAcceptedMessage>(MessageType.CONNECTION_ACCEPTED, { gameState: initialState, yourPlayerId: 'test' }),
       createMessage<ConnectionRejectedMessage>(MessageType.CONNECTION_REJECTED, { reason: 'test' }),
       createMessage<PingMessage>(MessageType.PING, {}),
