+++ src/games/farsketched/host/HostLobby.tsx
@@ -1,17 +1,13 @@
-import { useState, useEffect } from 'react';
-import { QRCodeSVG } from 'qrcode.react';
+import { useState } from 'react';
 import { Player, GameConfig } from '@/games/farsketched/types';
 import { usePeer } from '@/contexts/PeerContext';
 import {
   Box,
   Typography,
-  Button,
   Paper,
   Avatar,
-  Snackbar,
-  Alert,
 } from '@mui/material';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
+import { JoinGameQR } from '@/components/JoinGameQR';
 
 interface HostLobbyProps {
   gameConfig: GameConfig;
@@ -20,23 +16,6 @@ interface HostLobbyProps {
 
 export const HostLobby = ({ gameConfig, players }: HostLobbyProps) => {
   const { peerId } = usePeer();
-  const [joinUrl, setJoinUrl] = useState<string>('');
-  const [isCopied, setIsCopied] = useState(false);
-
-  useEffect(() => {
-    if (peerId) {
-      // const baseUrl = 'https://farsketched.netlify.app';
-      const baseUrl = 'http://localhost:8000'; 
-      const url = new URL('clientindex.html', baseUrl);
-      url.searchParams.set('roomCode', peerId);
-      setJoinUrl(url.toString());
-    }
-  }, [peerId]);
-
-  const handleCopyLink = () => {
-    navigator.clipboard.writeText(joinUrl);
-    setIsCopied(true);
-  };
 
   return (
     <Box sx={{ p: 3, height: '100%', maxWidth: '100vw', overflowX: 'auto' }}>
@@ -114,77 +93,8 @@ export const HostLobby = ({ gameConfig, players }: HostLobbyProps) => {
         </Paper>
 
         {/* QR Section */}
-        <Paper 
-          elevation={3} 
-          sx={{ 
-            p: 3, 
-            height: '100%',
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            bgcolor: 'rgba(255, 255, 255, 0.1)',
-            backdropFilter: 'blur(8px)',
-            minWidth: 0,
-            maxWidth: '100%',
-          }}
-        >
-          <Typography variant="h4" gutterBottom>
-            Scan to Join
-          </Typography>
-          {joinUrl && (
-            <Box sx={{ mt: 2, textAlign: 'center' }}>
-              <Box
-                sx={{
-                  p: 0.5,
-                  bgcolor: 'white',
-                  borderRadius: 2,
-                  display: 'inline-block',
-                  maxWidth: '100%',
-                  minWidth: 0,
-                }}
-              >
-                <QRCodeSVG
-                  value={joinUrl}
-                  size={256}
-                  level="H"
-                  marginSize={4}
-                />
-              </Box>
-              <Box sx={{ mt: 3 }}>
-                <Button
-                  variant="contained"
-                  startIcon={<ContentCopyIcon />}
-                  onClick={handleCopyLink}
-                  size="large"
-                >
-                  Copy Link
-                </Button>
-              </Box>
-            </Box>
-          )}
-        </Paper>
+        {peerId && <JoinGameQR peerId={peerId} gameName="farsketched" />}
       </Box>
-
-      <Snackbar
-        open={isCopied}
-        autoHideDuration={2000}
-        onClose={() => setIsCopied(false)}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-      >
-        <Alert 
-          severity="success" 
-          variant="filled"
-          sx={{ 
-            bgcolor: '#00c853',
-            color: 'white',
-            '& .MuiAlert-icon': {
-              display: 'none'
-            }
-          }}
-        >
-          Link copied to clipboard!
-        </Alert>
-      </Snackbar>
     </Box>
   );
 };
